// lib/categories-enum.ts
export enum CategoriesEnum {
  Acting = "ACTING",
  AgedCare = "AGED_CARE",
  AgribusinessManagement = "AGRIBUSINESS_MANAGEMENT",
  Ai = "AI",
  Animation_3D = "ANIMATION_3D",
  Automotive = "AUTOMOTIVE",
  Baking = "BAKING",
  BankingManagement = "BANKING_MANAGEMENT",
  Beauty = "BEAUTY",
  BlockchainTech = "BLOCKCHAIN_TECH",
  Blog = "BLOG",
  BusinessAnalysis = "BUSINESS_ANALYSIS",
  BusinessManagement = "BUSINESS_MANAGEMENT",
  Carpentry = "CARPENTRY",
  CloudComputing = "CLOUD_COMPUTING",
  CommercialCooking = "COMMERCIAL_COOKING",
  ComputerNetworking = "COMPUTER_NETWORKING",
  Construction = "CONSTRUCTION",
  DataManagement = "DATA_MANAGEMENT",
  Dental = "DENTAL",
  Design = "DESIGN",
  DigitalMedia = "DIGITAL_MEDIA",
  Education = "EDUCATION",
  Events = "EVENTS",
  FashionDesign = "FASHION_DESIGN",
  FilmMedia = "FILM_MEDIA",
  Fitness = "FITNESS",
  GameProgramming = "GAME_PROGRAMMING",
  GraphicDesign = "GRAPHIC_DESIGN",
  Hairdressing = "HAIRDRESSING",
  Health = "HEALTH",
  Hospitality = "HOSPITALITY",
  HospitalityManagement = "HOSPITALITY_MANAGEMENT",
  Horticulture = "HORTICULTURE",
  HumanResources = "HUMAN_RESOURCES",
  InteriorDesign = "INTERIOR_DESIGN",
  InternationalBusiness = "INTERNATIONAL_BUSINESS",
  IntlHotelManagement = "INTL_HOTEL_MANAGEMENT",
  It = "IT",
  Landscape = "LANDSCAPE",
  LeadershipManagement = "LEADERSHIP_MANAGEMENT",
  MachineLearning = "MACHINE_LEARNING",
  Massage = "MASSAGE",
  Management = "MANAGEMENT",
  Medicine = "MEDICINE",
  Mental_Health = "MENTAL_HEALTH",
  MusicAudio = "MUSIC_AUDIO",
  Nursing = "NURSING",
  Nutrition = "NUTRITION",
  ProjectManagement = "PROJECT_MANAGEMENT",
  Public_Health = "PUBLIC_HEALTH",
  SocialmediaMarketing = "SOCIALMEDIA_MARKETING",
  SoftwareDevelopment = "SOFTWARE_DEVELOPMENT",
  Sport = "SPORT",
  SupplychainManagement = "SUPPLYCHAIN_MANAGEMENT",
  Telecommunication = "TELECOMMUNICATION",
  TravelTourism = "TRAVEL_TOURISM",
  Trades = "TRADES",
  Veterinary = "VETERINARY",
  WebDevelopment = "WEB_DEVELOPMENT",
  Yoga = "YOGA",
}